refactor(message-list): restructure UI components into atomic design layers
