+++ feature/mail/message/list/api/src/main/kotlin/net/thunderbird/feature/mail/message/list/ui/MessageListContract.kt
@@ -64,7 +64,6 @@ interface MessageListContract {
          *
          * @param state The current state of the message list to be rendered.
          * @param dispatchEvent A lambda function to be invoked when a user action or other UI event occurs.
-         * @param onEffect A lambda function to handle one-time side effects from the ViewModel.
          * @param modifier The modifier to be applied to the root container of the message list screen.
          * @param inAppNotificationEventFilter A filter to decide whether an in-app notification should be displayed.
          */
@@ -72,7 +71,6 @@ interface MessageListContract {
         fun Render(
             state: MessageListState,
             dispatchEvent: (MessageListEvent) -> Unit,
-            onEffect: (MessageListEffect) -> Unit,
             modifier: Modifier = Modifier,
             inAppNotificationEventFilter: (InAppNotification) -> Boolean = { true },
         )
@@ -96,7 +94,7 @@ interface MessageListContract {
             inAppNotificationEventFilter: (InAppNotification) -> Boolean = { true },
         ) {
             val (state, dispatchEvent) = viewModel.observe(onEffect)
-            Render(state.value, dispatchEvent, onEffect, modifier, inAppNotificationEventFilter)
+            Render(state.value, dispatchEvent, modifier, inAppNotificationEventFilter)
         }
     }
 }
